Refactor permissions: split out phase and project settings
