+++ front/app/containers/Admin/projects/routes.tsx
@@ -33,7 +33,10 @@ const AdminProjectsProjectGeneral = lazy(() => import('./project/general'));
 const AdminPhaseNewAndEdit = lazy(() => import('./project/timeline/edit'));
 const AdminProjectEvents = lazy(() => import('./project/events'));
 const AdminProjectEventsEdit = lazy(() => import('./project/events/edit'));
-const AdminProjectPermissions = lazy(() => import('./project/permissions'));
+const AdminProjectPermissions = lazy(
+  () => import('./project/permissions/Project')
+);
+const AdminPhasePermissions = lazy(() => import('./project/permissions/Phase'));
 const AdminProjectSurveyResults = lazy(() => import('./project/surveyResults'));
 const AdminProjectPoll = lazy(() => import('./project/poll'));
 const AdminProjectsSurvey = lazy(() => import('./project/nativeSurvey'));
@@ -425,7 +428,7 @@ const createAdminProjectsRoutes = () => {
                 path: projectsRoutes.projectPhaseAccessRights,
                 element: (
                   <PageLoading>
-                    <AdminProjectPermissions />
+                    <AdminPhasePermissions />
                   </PageLoading>
                 ),
               },
